only register user after permissions done
Instead of the login function being called when the user consents on ConsentPage, we should login only once the permissions are sorted and we are on the SaveQrPage.
This prevents the possibility of so-called 'ghost users' who downloaded the app, consented, but quit before giving permissions. We'd prefer to not have a buildup of these empty opcodes registered on the server.

To do this we have a new flag in onboardingHelper, 'registerUserDone', and we listen in SaveQrPage for appStatus to be true. If it is, and we haven't already registerUserDone, we call the login function to register.
